tools: merge_dtbs: Respect miboard-id while merging
On qualcomm 5.10 devices the devicetree is built from multiple techpack repositories as multiple dtb(o)s which will be merged later. Xiaomi has introduced the xiaomi,miboard-id property to mark split devicetrees which are (in)compatible to merge. The public qualcomm script to merge dtbs does not respect this property and merges incompatible split devicetrees together which results in a broken devictree. This can be fixed by respecting the miboard-id in the merge dts script.

Change-Id: I2cde9aa3ce9a9dbfa55768c665cea8885ba2c6ea